ReadSim: genome length into long

+++ src/simulator/Sim_reads.cpp
@@ -227,11 +227,11 @@ void simulate_reads(std::string genome_file, std::string error_profile_file, int
 //	cout << "\t\tReads simulated: " << prev << "%" << std::endl;
 
 
-	std::map<std::string, int> cov_reported;
+	std::map<std::string, long> cov_reported;
 
 	for (std::map<std::string, std::string>::iterator i = genome.begin(); i != genome.end(); i++) {
 
-		cov_reported[(*i).first]=((int)(*i).second.size())*coverage;
+		cov_reported[(*i).first]=((long)(*i).second.size())*coverage;
 	}
 
 	std::map<std::string, std::string>::iterator current_chr = genome.begin(); // we just go one chr after another.
